# swagger/swagger-test.json
{
  "swagger": "2.0",
  "info": {
    "title": "N빵 API",
    "description": "SWM-ChocoBread N빵 API 명세서",
    "version": "1.0.0"
  },
  "host": "localhost:5005",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "http",
      "scheme": "bearer",
      "in": "header",
      "bearerFormat": "JWT"
    }
  },
  "paths": {
...
    "/deals/{dealId}": {
      "get": {
        "tags": [
          "Deals"
        ],
        "summary": "거래 세부정보 GET",
        "description": "",
        "parameters": [
          {
            "name": "dealId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      },
      "put": {
        "tags": [
          "Deals"
        ],
        "summary": "거래 수정",
        "description": "",
        "parameters": [
          {
            "name": "dealId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "content": {
                  "example": "any"
                },
                "totalPrice": {
                  "example": "any"
                },
                "personalPrice": {
                  "example": "any"
                },
                "totalMember": {
                  "example": "any"
                },
                "dealDate": {
                  "example": "any"
                },
                "dealPlace": {
                  "example": "any"
                },
                "currentMember": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      },
      "delete": {
        "tags": [
          "Deals"
        ],
        "summary": "거래 삭제",
        "description": "",
        "parameters": [
          {
            "name": "dealId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
...
    "/deals/{dealId}/users/{userId}": {
      "get": {
        "tags": [
          "Deals"
        ],
        "summary": "거래 유저 상태(참여자, 제안자, 참여하지 않음)",
        "description": "",
        "parameters": [
          {
            "name": "dealId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
...
    "/users/deals/{userId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "마이페이지 거래내역 조회",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/users/location/{userId}/{latitude}/{longitude}": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "네이버 GeoLocation으로 현 위치 저장",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "latitude",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "longitude",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/users/location": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "X-FORWARDED-FOR",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {},
        "deprecated": true
      }
    },
    "/users/{userId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "유저 정보 반환",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "닉네임 변경",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nick": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/users/kakaosdk/{kakaoNumber}": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "카카오 유저 닉네임 변경",
        "description": "",
        "parameters": [
          {
            "name": "kakaoNumber",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nick": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/users/check/{userId}/{nick}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "닉네임 중복 확인",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "nick",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/users/report/{userId}": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "유저 신고",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "content": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/users/check/{userId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "회원가입 여부 반환",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}